Update Models - TotalProductSales

+++ DBTeamWorkSidecar/Supermarket.Models/TotalProductSales.cs
@@ -14,7 +14,7 @@ namespace Supermarket.Models
         public int ProductId { get; set; }
         public string ProductName { get; set; }
         public string VendorName { get; set; }
-        public float QuantitySold { get; set; }
-        public decimal TotalIncomes { get; set; }
+        public double QuantitySold { get; set; }
+        public double TotalIncomes { get; set; }
     }
 }
